+++ ErsatzTV.Application/MediaCards/Mapper.cs
@@ -131,7 +131,7 @@ internal static class Mapper
         return new SongCardViewModel(
             songMetadata.SongId,
             songMetadata.Title,
-            songMetadata.Artist + album,
+            string.Join(", ", songMetadata.Artists) + album,
             songMetadata.SortTitle,
             GetThumbnail(songMetadata, None, None),
             songMetadata.Song.State);