+++ ErsatzTV.Core/Domain/Metadata/SongMetadata.cs
@@ -3,10 +3,10 @@
 public class SongMetadata : Metadata
 {
     public string Album { get; set; }
-    public string Artist { get; set; }
-    public string AlbumArtist { get; set; }
-    public string Date { get; set; }
+    public IList<string> Artists { get; set; }
+    public IList<string> AlbumArtists { get; set; }
     public string Track { get; set; }
+    public string Comment { get; set; }
     public int SongId { get; set; }
     public Song Song { get; set; }
 }