+++ ErsatzTV.Core/Interfaces/Metadata/ILocalStatisticsProvider.cs
@@ -9,5 +9,5 @@ public interface ILocalStatisticsProvider
     
     Task<Either<BaseError, bool>> RefreshStatistics(string ffmpegPath, string ffprobePath, MediaItem mediaItem);
 
-    Task<Either<BaseError, Dictionary<string, string>>> GetSongTags(string ffprobePath, MediaItem mediaItem);
+    Either<BaseError, List<SongTag>> GetSongTags(string ffprobePath, MediaItem mediaItem);
 }