+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
     - The default templates will be extracted and overwritten every time ErsatzTV is started
     - The templates use [scribian](https://github.com/scriban/scriban/tree/master/doc) template syntax
     - The templates contain comments describing which fields are available for use in the templates
+- Add `comment` to song metadata
 
 ### Fixed
 - Fix antiforgery error caused by reusing existing browser tabs across docker container restarts